requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
asyncpg==0.30.0
attrs==24.3.0
build==1.0.3
CacheControl==0.14.0
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.3.0
cleo==2.1.0
crashtest==0.4.1
cryptography==41.0.4
distlib==0.3.8
dulwich==0.21.6
fastjsonschema==2.18.1
filelock==3.16.1
frozenlist==1.5.0
greenlet==3.1.1
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
jaraco.classes==3.3.0
jeepney==0.8.0
keyring==24.2.0
magic-filter==1.0.12
Mako==1.3.8
MarkupSafe==3.0.2
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.1.0
packaging==24.1
pexpect==4.9.0
pipenv==2024.4.0
pkginfo==1.12.0
platformdirs==4.3.6
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
propcache==0.2.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pycparser==2.21
pydantic==2.10.5
pydantic_core==2.27.2
pyproject_hooks==1.0.0
python-dotenv==1.0.1
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.3
SQLAlchemy==2.0.37
tomlkit==0.12.1
trove-classifiers==2023.9.19
typing_extensions==4.12.2
urllib3==2.0.6
virtualenv==20.28.0
yarl==1.18.3
zipp==3.17.0

starlette~=0.45.2